Allow accessKeyId and secretAccessKey to be modified
Ran into an issue with NoSQL workbench where it generates it's own access and secret key. I couldn't use the `sharedDb` option the as we run dynamoDB from docker and it's shared with our unit testing.

Proposal to add an option so that these keys can be added in and customised so we can manage our DB through NoSQL workbench. I don't see this as a breaking change.